Added support for BWA style XA tags for counting.
These changes allows us to use bwa as the aligner under conditions
with multiply mapped reads.  BWA uses the XA flag to denote
additional alignments and sets the aligntment quality to 0.
This confuses htseq-count which disregards any reads with aqual0
by default.  This includes alignments that have the XA flag
in the counting.

t wraps the XA check in a try block.  This was done as the existing
code didn't use `.default` but instead caught the KeyError.  This
code detects of either read pair is None and if not checks for the
existence of an optional_field.